Diagnose more errors

+++ src/time.hh
@@ -6,7 +6,7 @@
 template<class Clock>
 int24_t get_time() {
     constexpr auto ONE_DAY = std::chrono::duration_cast<typename Clock::duration>(std::chrono::seconds(24 * 60 * 60));
-    constexpr auto TICKS_PER_UNIT = ONE_DAY.count() / static_cast<long double>(INT24_MAX);
+    constexpr auto TICKS_PER_UNIT = static_cast<long double>(ONE_DAY.count()) / static_cast<long double>(INT24_MAX);
 
     auto ticks = (Clock::now().time_since_epoch() % ONE_DAY).count();
     auto time = static_cast<long double>(ticks) / TICKS_PER_UNIT;
